Refactor Orchestrator types into multiple files
Single file was starting to become too big to find things in. I've
re-exported everything so that we don't have to move all the imports.
